Update stac from landsat bucket for different missions
When looking up a scene's metadata, if the STAC catalog did not have an
item for a scene, we'd fall back to looking for a STAC item JSON in the
AWS S3 bucket. This, however, doesn't work for legacy missions (L4,5,7)
as the bucket is organized by sensor and we hard-coded the L8,9 sensor.

Example scene: LE07_L1GT_063018_20170915_20200830_02_T2

This add the sensor mapping and dynamic sensor value in S3 path.